% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup1NE.R
\name{ETBD_migrateSYM.NE}
\alias{ETBD_migrateSYM.NE}
\title{ETBD_migrateSYM.NE.}
\usage{
ETBD_migrateSYM.NE(
  initialtree,
  t = 10,
  JmaxV = 1000,
  split = F,
  bud = T,
  siteN = 1,
  DIST = "SRS",
  psymp = 0.1,
  SADmarg = 0.1,
  exparm = -0.7,
  NegExpEx = T,
  ExpSpParm = 0.7,
  ExpSpParm2 = -0.1,
  ExpSp = T,
  splitparm = 0.5,
  constantEX = 0.1,
  migprob1 = 0,
  migprob2 = 0,
  exparm2 = 0.5,
  Asteroid = c(1000, 1800),
  Asteroidimpact = 0,
  GROW = F,
  initialsize = 100,
  threshold = 0.01,
  reggie = JmaxV,
  spud = 1
)
}
\arguments{
\item{t}{Number of time steps}

\item{JmaxV}{Resource levels (Jmax)}

\item{DIST}{Type of species abundance distribution: either log-normal: "NORM", Fisher's log-series: "SRS", or uniform:"NO" (SAD)}

\item{exparm}{Extinction parameter (x0)}

\item{NegExpEx}{If true indicates extinction is population size dependent and will utilize exparm paramaters.}

\item{ExpSpParm}{Speciation parameter (v0)}

\item{ExpSpParm2}{Speciation parameter (v1)}

\item{ExpSp}{If true indicates speciation is population size dependent and will utilize ExpSpParm paramaters.}

\item{splitparm}{Heritability of population size (h)}

\item{constantEX}{Constant probability of either speciation or extinction given one of the dependencies is false (con)}

\item{exparm2}{Extinction parameter (x1)}

\item{initialsize}{Initial population size of starting species}
}
\value{
A list containing:
\itemize{
  \item \code{tree} – The final newick tree  after simulation.
  \item \code{trees} – All trees at each time step
  \item \code{matrix_list} – A list of all species identities and population sizes for final tree
  \item \code{mig} – A list of all species identities and population sizes for every tree at each time step
}
}
\description{
Simulates phylogenetic trees under the assumptions of the Equilibrium Theory of Biodiversity Dynamics.
}
\examples{
ETBD_migrateSYM.NE(t =20,DIST = "NORM",JmaxV = 3807,NegExpEx = T,  exparm = -0.82, exparm2 = -0.70,psymp = 0.36, ExpSp = T,    ExpSpParm2 = -0.08,ExpSpParm = 0.94,constantEX =  0.36,splitparm = .5, )
}
